src/main/java/org/wecancodeit/evercraft/Hero.java
package org.wecancodeit.evercraft;

public class Hero {

    private String name;
    private Alignment alignment;

    public Hero(String name, Alignment alignment) {
        this.name = name;
        this.alignment = alignment;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Alignment getAlignment() {
        return alignment;
    }

    public void setAlignment(Alignment alignment) {
        this.alignment = alignment;
    }
}
